Added Meta Viewport Tag
Most Sites would look bad at 320 pixels wide.
So mobile browsers had a clever way to workaround this problem.  They
opened the page in a viewport and requested a window larger than the
screen size and zoomed out so the whole page would render.
While this solves the first problem, it creates another: the screen is
unreadable.
Adding this meta tag tells mobile browsers to size the viewport to the
device width.  It's a promise that you will be a responsible developer
that ensures the site will adjust appropriately at a small number of
pixels.