Added Content Container
Added div with `container` class.
Container is used as a parent level class to center a page's content at
various breakpoints.
http://getbootstrap.com/css/#overview-container

Added a style to the container and the body so it would be obvious.
Navbar adds a 20px bottom margin; cancelled it out so the container
wouldn't look so bad.