Created a horizontal form
You can save a little bit of space on large screen's by putting the
label and input control on the same line (not standard in Bootstrap
forms).

To do this, you add the `form-horizontal` class to the form.

When you do this, each container with class `form-group` behaves like a
row in a grid, so you can align the elements using column sizes.

The checkbox and submit button will need to be placed inside of a form
group so they get their own row

Since neither have a label to take up two columns, they'll both use an
offset.

http://getbootstrap.com/css/#forms-horizontal